Some storage formats that we support are not supported by overlayfs, so when we detect one disable that functionality.

+++ es-app/src/guis/GuiMenu.cpp
@@ -966,31 +966,38 @@ void GuiMenu::openSystemSettings_batocera()
 		auto mount_games = std::make_shared<SwitchComponent>(mWindow);
 		mount_games->setState(MountGamesEnabled);
 		s->addWithLabel(_("AUTODETECT GAMES CARD"), mount_games);
-		s->addSaveFunc([mount_games] {
+		mount_games->setOnChangedCallback([this, s, mount_games] {
 			SystemConf::getInstance()->setBool("system.automount", mount_games->getState());
 			SystemConf::getInstance()->saveSystemConf();
 		});
 
+		if (Utils::FileSystem::exists("/storage/.overlay_supported") && MountGamesEnabled) 
+		{
+	                auto overlayState = std::make_shared<SwitchComponent>(mWindow);
+	                bool overlayStateEnabled = SystemConf::getInstance()->getBool("system.merged.storage");
+	                overlayState->setState(overlayStateEnabled);
+	                s->addWithLabel(_("ENABLE MERGED STORAGE"), overlayState);
+	                overlayState->setOnChangedCallback([this, s, overlayState] {
+	                        bool overlayStateEnabled = overlayState->getState();
+	                        if (overlayStateEnabled) {
+	                          runSystemCommand("/usr/bin/systemctl start storage-roms.mount", "", nullptr);
+	                        } else {
+	                          runSystemCommand("/usr/bin/systemctl stop storage-roms.mount", "", nullptr);
+	                        }
+				SystemConf::getInstance()->setBool("system.merged.storage", overlayState->getState());
+				SystemConf::getInstance()->saveSystemConf();
+			});
+		}
+
 		s->addEntry(_("EJECT MICROSD CARD"), false, [window] {
-			runSystemCommand("/usr/bin/systemctl stop storage-roms.mount; /usr/bin/umount /storage/games-mmc; /usr/bin/systemctl start storage-roms.mount", "", nullptr);
+			if (Utils::FileSystem::exists("/storage/.overlay_supported"))
+			{
+				runSystemCommand("/usr/bin/systemctl stop storage-roms.mount; /usr/bin/umount /storage/games-mmc; /usr/bin/systemctl start storage-roms.mount", "", nullptr);
+			} else {
+				runSystemCommand("/usr/bin/umount /storage/roms", "", nullptr);
+			}
 			window->pushGui(new GuiMsgBox(window, _("You may now remove the card.")));
 		});
-
-                auto overlayState = std::make_shared<SwitchComponent>(mWindow);
-                bool overlayStateEnabled = SystemConf::getInstance()->getBool("system.merged.storage");
-                overlayState->setState(overlayStateEnabled);
-                s->addWithLabel(_("ENABLE MERGED STORAGE"), overlayState);
-                overlayState->setOnChangedCallback([this, s, overlayState] {
-                        bool overlayStateEnabled = overlayState->getState();
-                        if (overlayStateEnabled) {
-                          runSystemCommand("/usr/bin/systemctl start storage-roms.mount", "", nullptr);
-                        } else {
-                          runSystemCommand("/usr/bin/systemctl stop storage-roms.mount", "", nullptr);
-                        }
-			SystemConf::getInstance()->setBool("system.merged.storage", overlayState->getState());
-			SystemConf::getInstance()->saveSystemConf();
-                });
-
 	}
 
 	s->addGroup(_("HARDWARE / PERFORMANCE"));
